Implemented the unimplemented methods in VOControllingDesc.
git-svn-id: https://open-delta.googlecode.com/svn/trunk@666 dd429c29-c832-59a0-e864-46e68368aa0b

+++ delta-editor/src/main/java/au/org/ala/delta/editor/slotfile/model/SlotFileDataSet.java
@@ -392,7 +392,7 @@ public class SlotFileDataSet extends AbstractObservableDataSet {
     	    VOControllingDesc ctlDesc = (VOControllingDesc)getVOP().getDescFromId(id);
     	    if (compare(charId, stateId, ctlDesc)) {
     	        List<Integer> ctlStates = ctlDesc.readStateIds();
-    	        ctlStates.remove(stateId);
+    	        ctlStates.remove((Integer)stateId);
  
     	        changeControllingStates(ctlDesc, ctlStates);
     	    }
